start to populate page-template with Employee objects
