+++ src/generate-site.js
@@ -1,0 +1,19 @@
+const fs = require('fs');
+
+const writeFile = fileContent => {
+    return new Promise((resolve, reject) => {
+        fs.writeFile('./dist/index.html', fileContent, err => {
+            if (err) {
+                reject(err);
+                return;
+            }
+
+            resolve({
+                ok: true,
+                message: "File created!"
+            });
+        });
+    });
+};
+
+module.exports = writeFile;
